Demo Auth
Simple 3 page React app to demo the protected API with cognito OAuth2 authentication using Google identity provider and JWT authorization to protect the api.  Consists of a public WelcomePage, with a navigation button to the LoginPage, and a ProtectedPage that displays a ingredients/search api call using a JWT auth header.